Move render_process tests to their own folder
